refs #4565 migrate config super user to database

+++ plugins/UsersManager/API.php
@@ -15,7 +15,6 @@ use Piwik\Access;
 use Piwik\Common;
 use Piwik\Config;
 use Piwik\Date;
-use Piwik\Db;
 use Piwik\Option;
 use Piwik\Piwik;
 use Piwik\Site;
@@ -36,10 +35,20 @@ use Piwik\Tracker\Cache;
  */
 class API extends \Piwik\Plugin\API
 {
+    /**
+     * @var Model
+     */
+    private $model;
+
     const PREFERENCE_DEFAULT_REPORT = 'defaultReport';
     const PREFERENCE_DEFAULT_REPORT_DATE = 'defaultReportDate';
 
     static private $instance = null;
+
+    protected function __construct()
+    {
+        $this->model = new Model();
+    }
 
     /**
      * You can create your own Users Plugin to override this class.
@@ -125,18 +134,12 @@ class API extends \Piwik\Plugin\API
     {
         Piwik::checkUserHasSomeAdminAccess();
 
-        $where = '';
-        $bind = array();
+        $logins = array();
         if (!empty($userLogins)) {
-            $userLogins = explode(',', $userLogins);
-            $where = 'WHERE login IN (' . Common::getSqlStringFieldsArray($userLogins) . ')';
-            $bind = $userLogins;
+            $logins = explode(',', $userLogins);
         }
-        $db = Db::get();
-        $users = $db->fetchAll("SELECT *
-								FROM " . Common::prefixTable("user") . "
-								$where
-								ORDER BY login ASC", $bind);
+
+        $users = $this->model->getUsers($logins);
         // Non Super user can only access login & alias
         if (!Piwik::hasUserSuperUserAccess()) {
             foreach ($users as &$user) {
@@ -155,15 +158,7 @@ class API extends \Piwik\Plugin\API
     {
         Piwik::checkUserHasSomeAdminAccess();
 
-        $db = Db::get();
-        $users = $db->fetchAll("SELECT login
-								FROM " . Common::prefixTable("user") . "
-								ORDER BY login ASC");
-        $return = array();
-        foreach ($users as $login) {
-            $return[] = $login['login'];
-        }
-        return $return;
+        return $this->model->getUsersLogin();
     }
 
     /**
@@ -186,16 +181,7 @@ class API extends \Piwik\Plugin\API
 
         $this->checkAccessType($access);
 
-        $db = Db::get();
-        $users = $db->fetchAll("SELECT login,idsite
-								FROM " . Common::prefixTable("access")
-            . " WHERE access = ?
-								ORDER BY login, idsite", $access);
-        $return = array();
-        foreach ($users as $user) {
-            $return[$user['login']][] = $user['idsite'];
-        }
-        return $return;
+        return $this->model->getUsersSitesFromAccess($access);
     }
 
     /**
@@ -217,15 +203,7 @@ class API extends \Piwik\Plugin\API
     {
         Piwik::checkUserHasAdminAccess($idSite);
 
-        $db = Db::get();
-        $users = $db->fetchAll("SELECT login,access
-								FROM " . Common::prefixTable("access")
-            . " WHERE idsite = ?", $idSite);
-        $return = array();
-        foreach ($users as $user) {
-            $return[$user['login']] = $user['access'];
-        }
-        return $return;
+        return $this->model->getUsersAccessFromSite($idSite);
     }
 
     public function getUsersWithSiteAccess($idSite, $access)
@@ -233,17 +211,12 @@ class API extends \Piwik\Plugin\API
         Piwik::checkUserHasAdminAccess($idSite);
         $this->checkAccessType($access);
 
-        $db = Db::get();
-        $users = $db->fetchAll("SELECT login
-								FROM " . Common::prefixTable("access")
-            . " WHERE idsite = ? AND access = ?", array($idSite, $access));
-        $logins = array();
-        foreach ($users as $user) {
-            $logins[] = $user['login'];
-        }
+        $logins = $this->model->getUsersLoginWithSiteAccess($idSite, $access);
+
         if (empty($logins)) {
             return array();
         }
+
         $logins = implode(',', $logins);
         return $this->getUsers($logins);
     }
@@ -268,20 +241,9 @@ class API extends \Piwik\Plugin\API
     {
         Piwik::checkUserHasSuperUserAccess();
         $this->checkUserExists($userLogin);
-        $this->checkUserIsNotConfigSuperUser($userLogin);
+        $this->checkUserHasNotSuperUserAccess($userLogin);
 
-        $db = Db::get();
-        $users = $db->fetchAll("SELECT idsite,access
-								FROM " . Common::prefixTable("access")
-            . " WHERE login = ?", $userLogin);
-        $return = array();
-        foreach ($users as $user) {
-            $return[] = array(
-                'site'   => $user['idsite'],
-                'access' => $user['access'],
-            );
-        }
-        return $return;
+        return $this->model->getSitesAccessFromUser($userLogin);
     }
 
     /**
@@ -295,13 +257,8 @@ class API extends \Piwik\Plugin\API
     {
         Piwik::checkUserHasSuperUserAccessOrIsTheUser($userLogin);
         $this->checkUserExists($userLogin);
-        $this->checkUserIsNotConfigSuperUser($userLogin);
 
-        $db = Db::get();
-        $user = $db->fetchRow("SELECT *
-								FROM " . Common::prefixTable("user")
-            . " WHERE login = ?", $userLogin);
-        return $user;
+        return $this->model->getUser($userLogin);
     }
 
     /**
@@ -316,11 +273,7 @@ class API extends \Piwik\Plugin\API
         Piwik::checkUserHasSuperUserAccess();
         $this->checkUserEmailExists($userEmail);
 
-        $db = Db::get();
-        $user = $db->fetchRow("SELECT *
-								FROM " . Common::prefixTable("user")
-            . " WHERE email = ?", $userEmail);
-        return $user;
+        return $this->model->getUserByEmail($userEmail);
     }
 
     private function checkLogin($userLogin)
@@ -371,7 +324,6 @@ class API extends \Piwik\Plugin\API
         Piwik::checkUserHasSuperUserAccess();
 
         $this->checkLogin($userLogin);
-        $this->checkUserIsNotConfigSuperUser($userLogin);
         $this->checkEmail($email);
 
         $password = Common::unsanitizeInputValue($password);
@@ -382,17 +334,7 @@ class API extends \Piwik\Plugin\API
 
         $token_auth = $this->getTokenAuth($userLogin, $passwordTransformed);
 
-        $db = Db::get();
-
-        $db->insert(Common::prefixTable("user"), array(
-                                                      'login'           => $userLogin,
-                                                      'password'        => $passwordTransformed,
-                                                      'alias'           => $alias,
-                                                      'email'           => $email,
-                                                      'token_auth'      => $token_auth,
-                                                      'date_registered' => Date::now()->getDatetime()
-                                                 )
-        );
+        $this->model->addUser($userLogin, $passwordTransformed, $email, $alias, $token_auth, Date::now()->getDatetime());
 
         // we reload the access list which doesn't yet take in consideration this new user
         Access::getInstance()->reloadAccess();
@@ -412,32 +354,20 @@ class API extends \Piwik\Plugin\API
         $this->checkUserIsNotAnonymous($userLogin);
         $this->checkUserExists($userLogin);
 
-        $this->deleteUserAccess($userLogin);
-
-        $db = Db::get();
-        $db->update(Common::prefixTable("user"),
-            array(
-                'superuser_access' => $hasSuperUserAccess ? 1 : 0
-            ),
-            "login = '$userLogin'"
-        );
-    }
-
-    public function getUsersLoginHavingSuperUserAccess()
-    {
-        Piwik::checkUserHasSuperUserAccess();
-
-        $db = Db::get();
-        $users = $db->fetchAll("SELECT login
-                                FROM " . Common::prefixTable("user") . "
-                                WHERE superuser_access = 1
-                                ORDER BY login ASC");
-        $userLogins = array();
-        foreach ($users as $login) {
-            $userLogins[] = $login['login'];
+        if (!$hasSuperUserAccess && $this->isUserTheOnlyUserHavingSuperUserAccess($userLogin)) {
+            throw new Exception('You cannot remove Super User access from this user as there has to be at least one user having Super User access. You have to grant Super User access to another user first.');
         }
 
-        return $userLogins;
+        $this->model->deleteUserAccess($userLogin);
+
+        $this->model->setSuperUserAccess($userLogin, $hasSuperUserAccess);
+    }
+
+    public function getUsersHavingSuperUserAccess()
+    {
+        Piwik::checkUserIsNotAnonymous();
+
+        return $this->model->getUsersHavingSuperUserAccess();
     }
 
     /**
@@ -453,7 +383,6 @@ class API extends \Piwik\Plugin\API
     {
         Piwik::checkUserHasSuperUserAccessOrIsTheUser($userLogin);
         $this->checkUserIsNotAnonymous($userLogin);
-        $this->checkUserIsNotConfigSuperUser($userLogin);
         $userInfo = $this->getUser($userLogin);
 
         if (empty($password)) {
@@ -481,17 +410,8 @@ class API extends \Piwik\Plugin\API
         $alias = $this->getCleanAlias($alias, $userLogin);
         $token_auth = $this->getTokenAuth($userLogin, $password);
 
-        $db = Db::get();
+        $this->model->updateUser($userLogin, $password, $email, $alias, $token_auth);
 
-        $db->update(Common::prefixTable("user"),
-            array(
-                 'password'   => $password,
-                 'alias'      => $alias,
-                 'email'      => $email,
-                 'token_auth' => $token_auth
-            ),
-            "login = '$userLogin'"
-        );
         Cache::deleteTrackerCache();
 
         /**
@@ -515,13 +435,17 @@ class API extends \Piwik\Plugin\API
     {
         Piwik::checkUserHasSuperUserAccess();
         $this->checkUserIsNotAnonymous($userLogin);
-        $this->checkUserIsNotConfigSuperUser($userLogin);
         if (!$this->userExists($userLogin)) {
             throw new Exception(Piwik::translate("UsersManager_ExceptionDeleteDoesNotExist", $userLogin));
         }
 
-        $this->deleteUserOnly($userLogin);
-        $this->deleteUserAccess($userLogin);
+        if ($this->isUserTheOnlyUserHavingSuperUserAccess($userLogin)) {
+            throw new Exception('You cannot delete this user as there has to be at least one user having Super User access. To remove this user grant Super User access to another user and then you can delete this user.');
+        }
+
+        $this->model->deleteUserOnly($userLogin);
+        $this->model->deleteUserAccess($userLogin);
+
         Cache::deleteTrackerCache();
     }
 
@@ -539,10 +463,7 @@ class API extends \Piwik\Plugin\API
         Piwik::checkUserIsNotAnonymous();
         Piwik::checkUserHasSomeViewAccess();
 
-        $count = Db::fetchOne("SELECT count(*)
-                               FROM " . Common::prefixTable("user") . "
-                               WHERE login = ?", $userLogin);
-        return $count != 0;
+        return $this->model->userExists($userLogin);
     }
 
     /**
@@ -554,11 +475,8 @@ class API extends \Piwik\Plugin\API
     public function userEmailExists($userEmail)
     {
         Piwik::checkUserIsNotAnonymous();
-        $count = Db::fetchOne("SELECT count(*)
-								FROM " . Common::prefixTable("user") . "
-								WHERE email = ?", $userEmail);
-        return $count != 0
-        || Piwik::getConfigSuperUserEmail() == $userEmail;
+
+        return $this->model->userEmailExists($userEmail);
     }
 
     /**
@@ -605,21 +523,12 @@ class API extends \Piwik\Plugin\API
         // basically an admin can give the view or the admin access to any user for the websites he manages
         Piwik::checkUserHasAdminAccess($idSites);
 
-        $this->deleteUserAccess($userLogin, $idSites);
-
-        // delete UserAccess
-        $db = Db::get();
+        $this->model->deleteUserAccess($userLogin, $idSites);
 
         // if the access is noaccess then we don't save it as this is the default value
         // when no access are specified
         if ($access != 'noaccess') {
-            foreach ($idSites as $idsite) {
-                $db->insert(Common::prefixTable("access"),
-                    array("idsite" => $idsite,
-                          "login"  => $userLogin,
-                          "access" => $access)
-                );
-            }
+            $this->model->addUserAccess($userLogin, $access, $idSites);
         }
 
         // we reload the access list which doesn't yet take in consideration this new user access
@@ -667,13 +576,6 @@ class API extends \Piwik\Plugin\API
         }
     }
 
-    private function checkUserIsNotConfigSuperUser($userLogin)
-    {
-        if ($userLogin == Piwik::getConfigSuperUserLogin()) {
-            throw new Exception(Piwik::translate("UsersManager_ExceptionConfigSuperUser"));
-        }
-    }
-
     private function checkAccessType($access)
     {
         $accessList = Access::getListAccess();
@@ -686,53 +588,11 @@ class API extends \Piwik\Plugin\API
         }
     }
 
-    /**
-     * Delete a user given its login.
-     * The user's access are not deleted.
-     *
-     * @param string $userLogin the user login.
-     */
-    private function deleteUserOnly($userLogin)
+    private function isUserTheOnlyUserHavingSuperUserAccess($userLogin)
     {
-        $db = Db::get();
-        $db->query("DELETE FROM " . Common::prefixTable("user") . " WHERE login = ?", $userLogin);
+        $superUsers = $this->getUsersHavingSuperUserAccess();
 
-        /**
-         * Triggered after a user has been deleted.
-         * 
-         * This event should be used to clean up any data that is related to the now deleted user.
-         * The **Dashboard** plugin, for example, uses this event to remove the user's dashboards.
-         * 
-         * @param string $userLogin The login handle of the deleted user.
-         */
-        Piwik::postEvent('UsersManager.deleteUser', array($userLogin));
-    }
-
-    /**
-     * Delete the user access for the given websites.
-     * The array of idsite must be either null OR the values must have been checked before for their validity!
-     *
-     * @param string $userLogin the user login
-     * @param array $idSites array of idsites on which to delete the access. If null then delete all the access for this user.
-     *
-     * @return bool true on success
-     */
-    private function deleteUserAccess($userLogin, $idSites = null)
-    {
-        $db = Db::get();
-
-        if (is_null($idSites)) {
-            $db->query("DELETE FROM " . Common::prefixTable("access") .
-                " WHERE login = ?",
-                array($userLogin));
-        } else {
-            foreach ($idSites as $idsite) {
-                $db->query("DELETE FROM " . Common::prefixTable("access") .
-                    " WHERE idsite = ? AND login = ?",
-                    array($idsite, $userLogin)
-                );
-            }
-        }
+        return 1 >= count($superUsers) && Piwik::hasTheUserSuperUserAccess($userLogin);
     }
 
     /**
